Clarify comment and fix types for ContainsRange

+++ gocroaring.go
@@ -133,8 +133,8 @@ func (rb *Bitmap) Contains(x uint32) bool {
 	return answer
 }
 
-// ContainsRange returns true if the integer range is contained in the bitmap
-func (rb *Bitmap) ContainsRange(x, y uint32) bool {
+// ContainsRange returns true if the integers in the range [x, y) are contained in the bitmap
+func (rb *Bitmap) ContainsRange(x, y uint64) bool {
 	answer := bool(C.roaring_bitmap_contains_range(rb.cpointer, C.uint64_t(x), C.uint64_t(y)))
 	runtime.KeepAlive(rb)
 	return answer
